refactor: 패common ¤지 구조 변경
