+++ src/main/java/com/programmers/ticketparis/exception/BusinessException.java
@@ -1,4 +1,4 @@
-package com.programmers.ticketparis.common.exception;
+package com.programmers.ticketparis.exception;
 
 import java.util.List;
